Fix install problems
run.sh was not working outside virtual environment as it did not
declared explicitly the python path to be used
(oneview-redfish-toolkit root folder)

Also, using the launched version, it was noticed that pbr lib was
mistakenly considering this service as a lib that should be
installed using pip thus having a version number associated to it
that could not be loaded at runtime.

We also reverted the default security option to adhoc so that a
new self signed certificate is generated by flask every time the
service is launched instead of disabling by default.