Allows create a Composed System without Storage properly
Before this commit when we are trying create a
Composed System without storage and the Server Profile
related have not a local storage controller configured
properly, we receive a error.
Now, we are allowing create a Composed System without
Storage when the Server Profile Template related has not
a valid local storage controller configured.